requirements.txt
Flask>=2.0.0
google-genai>=0.2.0
flask-cors>=3.0.0
python-dotenv>=0.19.0

